# Refactor the Hyde::path() helper to normalize already absolute paths