--- a/packages/framework/src/Console/Concerns/Command.php
+++ b/packages/framework/src/Console/Concerns/Command.php
@@ -84,7 +84,7 @@ abstract class Command extends BaseCommand
      */
     public static function fileLink(string $filepath, string $label = null): string
     {
-        $link = 'file://'.str_replace('\\', '/', realpath($filepath) ?: Hyde::path(Hyde::pathToRelative($filepath)));
+        $link = 'file://'.str_replace('\\', '/', realpath($filepath) ?: Hyde::path($filepath));
 
         return $label ? "<href=$link>$label</>" : $link;
     }