### Refactor the Hyde::path() helper to normalize already absolute paths

Originally there has been two reasons for now doing this:

1. Paths should almost always already be relative
2. Performance concerns due to increased operations

After actually measuring this, I can confidently state that concern two is no concern at all. Firstly, performance is not that big of a concern in a static site generator, and secondly: after benchmarking this, it's clear that the method is lightning fast in either case.

### Benchmarks

| Input string           | Avg time (before) | Avg time (after) |
|------------------------|-------------------|------------------|
| ''                     | 0.001ms           | 0.001ms          |
| 'foo'                  | 0.002ms           | 0.003ms          |
| Hyde::path('foo')      | 0.004ms           | 0.006ms          |
| Hyde::path(Hyde::path) | 0.003ms           | 0.004ms          |

*Average time over 10 000 iterations
**Speed is unaffected by length and complexity of the input string

### Conclusion 

As you can see, the method is lightning fast in either case and will never be a bottleneck or even be noticeable. Even though it's a very frequently called method, since we can call it over 150 times per millisecond (with worst case performance) this just isn't a concern. The only difference is that it now also works with absolute paths.

Then only the first issue remains, and I think the improved leniency and thus beter DX is preferable here, thus the refactor.